Allow usage of subdomains in saml
At consume endpoint we now redirect the user
to the exact same path and query params and cookies
but to a host that is the same as the target_url.
For it to be possible target url must be passed in a cookie
that is shared between subdomains.

This allows developer to register only one domain at
idp but the redirect the user to an appropriate one
when they consume.